feat: add technical support URL in the LTI-based provider proctored exam download instructions (#1258)
This commit changes the installed version of the frontend-lib-special-exams module from 2.25.0 to 2.26.0.

This release adds support for displaying a technical support URL for LTI-based providers on the download instructions interstitial of proctored exams. The download instructions will display a technical support URL when it is returned from the proctoring settings backend endpoint. If the technical support URL is not available, then the technical support email and technical support phone number will be used instead.