feat: update version of frontend-lib-learning-assistant to 2.19.2
This commit installs version 2.19.2 of @edx/frontend-lib-learning-assistant.

This release commit fixes a bug where the days remaining banner appears after an audit trial learner sends their first message. In this case, the days remaining is not displayed until the call to the chat summary endpoint completes. This commit adds a loading spinner to the banner that appears while that call is in progress.

See https://github.com/edx/frontend-lib-learning-assistant/releases/tag/v2.19.2.